refactor: 중복으로 사용하는 check function을 common 폴더로 분리
