+++ src/controllers/common/checkMoneyInput.js
@@ -1,0 +1,9 @@
+import { isPositiveInteger } from "./isPositiveInteger.js";
+
+const isValidMoney = money => {
+  return (
+    money !== "" && isPositiveInteger(money) && parseInt(money, 10) % 10 === 0
+  );
+};
+
+export { isValidMoney };
